Mac: fix up the actual executable name for inspector
Use the process executable instead of the entry assembly to infer
whether workbooks or inspector is running.

I had toyed with using cecil to change the assembly name (not just
rename it), but doing so breaks registration in the XM runtime, and
there is not a clean way to redirect it. We would have to re-mmp
the whole app, which is very slow.