Xamarin.Interactive.Client: appease VS Mac
The VS Mac project model is still not fully backed by MSBuild. As such,
it (a) ignores Conditions on Reference/ProjectReference items, and
(b) does not pick up on any Reference/ProjectReference in imported
targets files.

So the trick is to make the csproj look Mac-only, with proper conditions
that are ignored in the IDE (but respected in actual MSBuild) and import
windows-specific targets that both VS and MSBuild will respect on
Windows.