Remove XCB dependency in ClientSession, WorkbookPageView
This abstracts the web view loading dance that happened in ClientSession
into an IWorkbookPageViewHost. It separates actual model/view/eval code
from WorkbookPageView to become WorkbookPageViewModel and XcbWorkbookPageView.

It is now possible to write new views for rendering and interacting
with a workbook.

Further work to split things into more sensible pieces will follow
organically as we move towards new UI models and hosting environments.

For instance, it's strange that XcbWorkbookPageView is actually a
subclass of WorkbookPageViewModel, but I wanted to keep everything as
cleanly split yet unchanged as possible for this first pass.