LogProvider: allow setting output writer (and allow null)
Our LogProvider always wrote to stderr if not connected to a debugger.
This needs to be configurable for supporting a console client.

Specify no writer when initialized by Agent. When connected to a client,
its logs will be sent to the client and rendered by the client's
LogProvider, and will still be sent to Debug.WriteLine if the agent
itself is connected to a debugger.