Win: Add Workbooks.exe location registry key
We don't strictly need this right now but it could be helpful down the road
or for third parties.

There's no need to ever use the Inspector key in `InteractiveInstallation`.
In the near future, the Inspector client will use a different technique to
locate installed bits anyway
(https://github.com/Microsoft/workbooks/issues/23).