Convert Xamarin.Interactive.dll from PCL to .NET Standard
Additionally, the Xamarin.Interactive.Shared project that was consumed
by each agent assembly is gone. All those types are now part of the
actual Xamarin.Interactive.dll. Same with
Xamarin.Interactive.Bootstrap.

InspectorSupport is no longer a partial class, because the shared
projects are gone now. But we need to keep the same API for Inspector
extension compatibility, and because that API is based on static
methods there's no helpful way to abstract any of it. So there is a
little duplication now.

Xamarin.Interactive.Facades is no longer necessary, and neither is the
PCL loading code in `UnifiedInspectorSupport`.

Xamarin.Forms has been bumped from 2.3.4.231 to 2.4.0.18342 for .NET
Standard 2.0 support.

`CompilationWorkspaceFactory` has to do some manual loading of
`Xamarin.Interactive.dll` and `netstandard.dll` now, but this will go
away when https://github.com/dotnet/roslyn/issues/20920 gets fixed.