Client: Prevent NuGet deps from accidentally being marked explicit
Rearrange the `InteractivePackage.IsExplicit` filtering in
`InitializePackagesAsync` so that `WorkbookPage.Packages` also gets
filtered. The latter is empty when a workbook is unsaved, but can
contain package deps when the workbook is saved.

The repro for this bug was:

1. Create a .NET Core workbook.
2. Add `System.Net.Http`.
3. Save.
4. Switch to Mono.
5. Witness giant package list.

Also fix the redundant check in
`InteractivePackageManager.GetInteractivePackageFromReader` to be more
correct. It shouldn't get hit now, though.