Compilation: Don't put any spaces in the generated assembly name.
Most things work fine with the space in the generated assembly name, but
if it is passed to COM anywhere (notably, as a parameter to a custom
attribute constructor), a TypeLoadException is encountered because the
COM-side doesn't seem to be able to load the type (though the assembly
is resolved just fine!). Replace the ' XIS' part of the string with a
bear emoji (a roughly equivalent 4 byte sequence).