Compilation: Improve detection of assembly name when patching compilation
Part 2 of fixing custom assemblies! When a type defined inside of a submission
is provided to an attribute as a parameter, it's encoded as such:

```
.custom instance void '<monkeyface>0001#0-0'/MyCustomAttribute::.ctor(class
    [mscorlib]System.Type) = ( 01 00 10 F0 9F 90 B5 30 30 30 30 23 30 2D 30 2B
    46 6F 6F 0 0 )
```

The bytes there are a string bytes: a few header bytes, F0 9F 90 B5 is the
monkey face emoji, 30 30 30 30 is 0000, 2B is a +, and 46 6F 6F are the string
`Foo`, the type name, then two null bytes indicating the end of the string. This
is standard encoding for a nested type name, and classes defined in
submissions *are* nested in a broader class defined for the submission.

Notice that the `0001` indicates that we're in the _second_ submission, but the
type name hasn't been rewritten for the custom attribute. The crux here is that
because of the `+` that is used for nested types, the script compilation patcher
doesn't break early enough, and the matched string doesn't match the full
assembly name. By adding `+` to the compilation patcher's termination markers,
it allows it to find the assembly name correctly in the nested type definition,
and patch it as well.