Forms: Fix loading agent extension when creating workspace.
When inspecting a Forms app, we load the Xamarin.Forms agent extensions
while we're creating the compilation workspace. We need to pass the
current session's evaluation context ID, but we can't access it via
`clientSession.CompilationWorkspace`, as that's what we're building now
(the compilation workspace), and the property is null. It's available to
both callers of this method, so pass it in as a parameter instead of
trying to get it via the client session.