Reflection: Load native deps for Kestrel's Libuv transport.
Add another special case here. This one is slightly trickier because
there is no _managed_ Libuv DLL that is ever loaded--instead, it's the
Kestrel transport that we're keying off of for now.

This also means that we can't just pick any random Libuv NuGet package
and use that native version, as there might be API breakage. To ensure
we're on the right version, read the nuspec file, determine the best
match based on the depenency of the current package, and load that.