Evaluation: Clean up assembly lookup handling a bit.
Previously, we had seen two different formats for assembly load
requests when attempting to resolve submission assemblies: by name
only (ie. `AssemblyFoo`) requests and by fully qualified assembly
name (`AssemblyFoo, Version=1.0.0.0, ...`) requests. The former
was added as a fix for bug
https://bugzilla.xamarin.com/show_bug.cgi?id=58801. ASP.NET Core running
on .NET Core introduced a _3rd_ type of request: a
fully-qualified-except-for-the-version assembly name. This was not in
our lookup tables, so things like looking up the startup class for an
ASP.NET Core server would fail.

Unify the handling here by implementing a custom comparer that only
checks the assembly name in a case-insensitive way, as the assemblies that
end up being cached in this lookup table are _always_ submission
assemblies, never have any version other than 0.0.0.0, never
have a culture, and never have a public key token.