Telemetry: only submit the AppSessionStart telem event
This event object is the only thing that has proven to be of use
to us so simply filter out all other events whens submitting.

The plumbing remains in place for other events as they can be
of use locally when debugging and such.

Change the endpoint for submission as well since we are shutting
down the v1 endpoint.