AgentIdentity: move AgentType/ID mappings from WorkbookAppInstallation
We really need to fully obsolete AgentType in favor of the
flavor/profile IDs that were introduced with WorkbookAppInstallation
since that scheme allows for arbitrary workbook app / agent types
and AgentType itself is fixed.

For now, move the mappings out of WorkbookAppInstallation and into
AgentIdentity, adding a FlavorId property that for now uses the static
map, but in the future should change to a field.

Mark these new APIs as obsolete and opt-in to them where needed today.
This will prevent future usages from going unnoticed.