XamPub: new model and serialization
This represents the new model that the Xamarin publisher pipeline has
been using for a while now. The old format is obsolete and it's time
we move to the new one.

Provides the basic model and some nice serialization helpers around it
to help shape resulting JSON in nicer human readable formats.