Web: Red underline diagnostics as you type
UpdateCodeCell now returns a CodeCellStatus, which includes any
diagnostics to underline, as well as a guess at whether or not the
submission is "complete", meaning that Enter will submit instead of
inserting a newline (enter handling not implemented in this commit).

UpdateCodeCell is now driven from MonacoCellEditor, which calls it
during editor change notifications, instead of waiting for React updates
to propagate the info. This lets MonacoCellEditor get the new return
value easily, and probably just makes more sense for us anyway.