Client: NuGet references go in hidden cell
Instead of generating a new cell with #r references at the end of the
workbook (which can be easily broken by the user, and complicates the
web flow), we now maintain a "hidden" first code cell and reevaluate
the entire workbook after adding a NuGet package.

ClientSession drives this behavior, as before. IEvaluationService now
has `AddTopLevelReferencesAsync`. The implementation on the web side is
trivial. On the desktop side, `Cell`s have grown the concept of being
non-serializable, as hidden cells show up in the document model. They
are excluded from the view model, though.