Web: Package search/install on client
Implement package search and installation as a dialog. Unfortunately,
the Fabric CommandBar's search isn't really usable at this time, so
replace it with a button to summon the dialog.

The dialog can be dismissed by clicking outside of it (unless an
installation is in progress).

PackageSearch component uses several Fabric components, but we are
unable to use their List at this time, as cells don't seem to re-render
on state changes.

Packages that are already installed are marked as such, and cannot be
installed again.