HostEnvironment, ClientAppHostEnvironment: split
Rename HostEnvironment to ClientAppHostEnvironment and introduce a new
HostEnvironment static class inside of XI that surfaces an OS property
of the existing HostOS enum type.

Remove InteractiveInstallation.IsMac and move all usages of that to
HostEnvironment.OS.

This eliminates the spidering dependency on InteractiveInstallation
for unrelated reasons.

Move DevEnvironment from XIC to XI as well.