MainThread: allow null SynchronizationContext
If we do not have a SynchronizationContext against which delegates
should be run, dispatch them directly. In this case, also make
Ensure() log a warning instead of throwing an error.

In an ASP.NET Core hosted world, we do not have a sync context,
nor do we want one, which would be a big bottleneck. We will need to
audit any non-GUI code that wants to post to the sync context.