Client: explicitly depend on NJS 11, C# Latest
We were previously implicitly using NJS because it was brought in by
NuGet. We actually want the new proper netstandard2.0 NJS along with
some of its nice new features now, so bring that in explicitly.

Also no reason not to bump to C# 7.2 (actually, Latest, because live
dangerously). I want default.