IWorkspaceService: refactor compilation
This defines an interface through which all compilation and IntelliSense
features are exposed and fits the Roslyn workspace to it. This is the
first big step towards supporting F# and other non-Roslyn-based
languages.

The following concepts are now decoupled from Roslyn:

- Cell structure
- Cell buffers
- Diagnostics
- Buffer locations and spans
- Completion
- Signature Help
- Hover/tooltips

There is still some implicit dependency around some minor Roslyn
structures (namely SourceText), but these should be easy to eliminate
in a second pass which will include moving the Roslyn workspace into
its own assembly.