PackageManagerService: pull out from ClientSession
Similar to what was done with pulling evaluation out of WPVM, this
pulls all of the "public" API for NuGet package management out of
ClientSession into a new PackageManagerService class and adapts
ClientSession to use it. It also provides some DTO types that are
public facing which currently sit on top of some of our internal
types that had similar responsibilities but were too coupled to
NuGet itself.