AgentClient: bring back message channel awaiters
We had this concept a long time ago - the ability for an agent
request to block on a response delivered on the message channel.

It was removed because the behavior was not actually necessary
in the web UI since Evaluate was changed to no longer return a
result (evaluation results are posted out of band on the channel,
since there may be any number of them for a cell at any time, even
after evaluation has finished for the cell).

However, this is necessary from an API behavior perspective to
guarantee that the call to Evaluate will yield _after_ the real
evaluation result is posted to the channel. Without blocking on
a response on the message channel, the behavior is not as one
would expect, and this causes UI complications with linear UX
(e.g. console client).

Before:
  → Begin: await Evaluate():
    ← Evaluation started event
    ← Evaluation finished event
    End: await Evaluate()
  ← Evaluation result

After:
  → Begin: await Evaluate():
    ← Evaluation started event
    ← Evaluation finished event
    ← Evaluation result
    End: await Evaluate()