CodeAnalysis: port everything to Monaco DTO models
This is a large pass that moves almost everything away from a Roslyn
dependency and eliminates a number of intermediate view model allocs
and conversions.

It introduces the XI.CodeAnalysis.Models namespace which contain
pure immutable DTOs which happen to serialize directly to the models
that Monaco expects.

This moves us close to being able to split out Roslyn entirely from
the core.