Port all serialization to Newtonsoft.Json
This finally drops using BinaryFormatter for serialization in favor
of Newtonsoft.Json. We must take special care to only use the
[JsonConstructor] attribute on all types we wish to serialize, and
none of the other attributes to influence serialization. Because
Xamarin.Interactive now ILRepacks, renames, and internalizes all of
Newtonsoft.Json, any external use of Newtonsoft.Json will not see
annotations on types (such as Newtonsoft.Json by way of SignalR in
the web client).

All converters must be defined in InteractiveJsonSerializerSettings and
registered manually (e.g. never use [JsonConverter] on a type).

[JsonConstructor] is special cased in the serializer settings' contract.