Roslyn: fix IntelliSense controller regressions
The controllers relied on a SourceTextContainer that could map back
to a Document instance. Due to a previous change where our SourceText
and SourceTextContainer was eliminated, this assumption is no longer
correct.

Instead, refactor the controllers to take a Document instance directly
and fetch SourceText from that only when needed.

In CompletionController, perform a larger cleanup to remove all
instance state except the ModelComputation. Roll some methods into
lambdas or inner functions to improve locality.