TypeSpec: support full construction from System.Type
Previously TypeSpec from Type was constructed by calling Type.ToString
and then parsing the type spec string that was produced. This is
inefficient and lossy around open generics - the generic type name
is lost via Type.ToString when the type is not closed.

We can retain this for open generics by constructing the TypeSpec
directly from the Type.