WebAssembly: Add a WebAssembly workbook target to the Web client.
Builds on the Mono WebAssembly SDK to provide a workbook target that evaluates
workbooks through a version of Mono compiled for WebAssembly. Provides the same
basic set of functions as the Console and .NET Core agents (evaluation only, no
inspection, view highlighting, etc.).

NuGet is supported (libraries have not been extensively tested, but
Newtonsoft.Json and a few other common things work); the workbook claims to be
.NET Standard 2.0 for the purposes of NuGet support.

Performance is also not terrible--initializing the agent takes anywhere from a
second to 3-4 depending on the browser's WASM implementation. Evaluations are
not noticeably slower than any of the existing agents, either. In Debug mode,
agent initialization is significantly slower at least on Chrome--Firefox and
Chakra do not suffer from the same issues. Safari doesn't work at all, Mono
crashes the WASM JIT there.

There's currently no network support, because there is no support in the WASM
runtime for it. We could and should implement support for HttpClient by
piggybacking it onto the browser fetch mechanism like Blazor did. There are also
a few agent client bits that are not implemented here that could be/will need to
be as well:

    * Setting the agent log level
    * Aborting evaluations
    * Loading assemblies into the agent for integrations

These are marked with TODOs in the code.