Core: Preparatory work for WebAssembly agent
The WebAssembly agent will need a special ticket/client/ECM because there's no
actual active connection from the compilation server to the agent--prepare the
core so that workbook apps are able to specify an assembly that will register a
"ticket provider," exactly like the agent process manager registration. The
registered ticket type will be created instead, and will be passed all the same
parameters as the normal AgentProcessTicket constructor receives.