CodeAnalysis: Resolving: add Mono's DllMap support
This is a managed implementation of Mono's DLL mapping support that
allows for platform-specific mapping of native libraries and symbols
within them.

We will use this as a basis for providing better native library loading
support for NuGet in .NET Core in particular via AssemblyLoadContext
LoadUnmanagedDll(string unmanagedDllName).

Unfortunately AssemblyLoadContext does not support symbol mapping, but
it is implemented in the DllMap in this commit should the ability be
introduced in the future.