CodeAnalysis: Split EvaluationAssemblyContext for .NET vs .NET Core.
On .NET Core, AssemblyLoadContext is a new managed API exposing some of the load
context concepts that have existed in the underlying framework for ages.

Split the existing EvaluationAssemblyContext into a .NET EAC, which reuses the
existing code and hooks the AppDomain's assembly loading events, and a .NET Core
version which creates a custom AssemblyLoadContext that hooks the default
context to provide assembly loads for evaluation, and hooks unmanaged loading so
that we can provide correct paths to native dependencies.