Web: initial work for overhauled result/representation UI
This is still largely a work in progress, but it’s functional enough
and an improvement over what is in master currently.

It brings a cleaner UI to result rendering and a much simpler approach
for implementing renderers.

There is a new representation selector that gets out of the way
of the cell to help reduce vertical space waste. The run button is
also moved into the cell itself as well.