Evaluation: Listen for loads in EvaluationAssemblyContext as well.
Sometimes, an assembly that we've sent over from the client can be loaded
without necessarily going through the existing Resolve event--for example, if an
assembly is loading dependencies and the dependent assembly is in the load
path. In those cases, we may not load external dependencies even though we have
them. Listening for load events ensures that we load external dependencies for
any assembly where they're available.