build: scan all projects in git; use actual MSBuild
This fixes the issue where dependencies.json was always changing
because it was based on the solution that was being built. Since we
now generate a solution based on a build profile, the repo-wide
dependencies graph was often changing incorrectly.

Instead of scanning a solution, just scan all projects in the repo
(via git ls-files, which is a big speed up vs walking the file system
tree).

Also eliminates all of the lame MSBuild solution and project parser
and leverages actual MSBuild since we have depend on that now for
other things.