Fix exception filtering on .NET Framework and .NET Core
Captured traces is an internal detail in Mono, so the filtering
we had did not work on .NET Core or .NET Framework. Filter frames
based on name instead by finding the frame that represents the
"Main" (🐵*.<<Initialize>>) for a cell and discarding frames below it,
which are an implementation detail of Workbooks.