Client: Show correct error when unable to connect XF workbook to Android
If connection fails, correct "Unable to start Android emulator" message
gets pushed to the status area.

"Optional Features" configuration was ignoring connection status. It
then failed to set up the XF feature, and the error message for this
failure took precedence in the status area.

Have connection init return a boolean value indicating success so that
feature configuration work can be skipped when there is no connection.

The correct error message is now displayed to the user in this case.